+++ packages/better-define/tests/fixtures/issue-806.vue
@@ -1,0 +1,11 @@
+<script lang="ts" setup>
+interface Foo {
+  (bar: number): void
+  a?: string
+  new (): void
+}
+
+defineProps<{
+  foo?: Foo
+}>()
+</script>