+++ packages/common/src/unplugin.ts
@@ -2,7 +2,7 @@ import {
   createFilter as createRollupFilter,
   type FilterPattern,
 } from '@rollup/pluginutils'
-import { generateTransform } from 'magic-string-ast'
+import { generateTransform, type CodeTransform } from 'magic-string-ast'
 import {
   REGEX_SETUP_SFC,
   REGEX_SRC_FILE,
@@ -12,7 +12,7 @@ import {
 } from './constants'
 import type { ResolvedOptions } from '@vitejs/plugin-vue'
 import type { Plugin } from 'rollup'
-import type { Plugin as VitePlugin } from 'vite'
+import type { HmrContext, Plugin as VitePlugin } from 'vite'
 
 /** @deprecated use `generateTransform` instead */
 export const getTransformResult: typeof generateTransform = generateTransform
@@ -96,4 +96,22 @@ export function getFilterPattern(
     filter.push(REGEX_SRC_FILE)
   }
   return filter
+}
+
+export function hackViteHMR(
+  ctx: HmrContext,
+  filter: (id: string | unknown) => boolean,
+  callback: (
+    code: string,
+    id: string,
+  ) => CodeTransform | undefined | Promise<CodeTransform | undefined>,
+): void {
+  if (!filter(ctx.file)) return
+
+  const { read } = ctx
+  ctx.read = async () => {
+    const code = await read()
+    const result = await callback(code, ctx.file)
+    return result?.code || code
+  }
 }