+++ packages/common/tsup.config.ts
@@ -2,4 +2,5 @@ import { config } from '../../tsup.config.js'
 
 export default config({
   onlyIndex: true,
+  platform: 'node',
 })