+++ packages/export-props/tests/fixtures/basic.vue
@@ -4,5 +4,7 @@ export const bar = 10
 export let baz: string | undefined
 export var qux = fn()
 
+console.log(foo, bar, baz)
+
 function fn() {}
 </script>