+++ packages/volar/tsup.config.ts
@@ -1,4 +1,5 @@
 import { config } from '../../tsup.config.js'
 export default config({
   ignoreDeps: ['vue-tsc'],
+  platform: 'node',
 })